Support GHC 9.4
