+++ src/Calligraphy/Compat/Lib.hs
@@ -14,12 +14,14 @@ module Calligraphy.Compat.Lib
     showAnns,
     mergeSpans,
     isPointSpan,
+    getHieFiles,
   )
 where
 
 import Calligraphy.Util.Lens
 import Data.IORef
 import qualified Data.Set as Set
+import qualified Calligraphy.Compat.GHC as GHC
 
 #if MIN_VERSION_ghc(9,0,0)
 import GHC.Iface.Ext.Binary
@@ -28,6 +30,7 @@ import GHC.Types.Name.Cache
 import GHC.Types.SrcLoc
 import GHC.Utils.Outputable (ppr, showSDocUnsafe)
 import qualified Data.Map as Map
+import Control.Monad (forM, when)
 #else
 import HieBin
 import HieTypes
@@ -35,11 +38,49 @@ import NameCache
 import SrcLoc
 #endif
 
+getHieFiles :: [FilePath] -> IO [HieFile]
+#if MIN_VERSION_ghc(9,4,0)
+
+getHieFiles filePaths = do
+  ref <- newIORef =<< GHC.initNameCache 'z' []
+  forM filePaths (readHieFileWithWarning ref)
+
+#else
+
+getHieFiles filePaths = do
+    uniqSupply <- GHC.mkSplitUniqSupply 'z'
+    ref <- newIORef (GHC.initNameCache uniqSupply [])
+    forM hieFilePaths (readHieFileWithWarning ref)
+
+#endif
+
+readHieFileWithWarning :: IORef GHC.NameCache -> FilePath -> IO GHC.HieFile
+readHieFileWithWarning ref path = do
+  GHC.HieFileResult fileHieVersion fileGHCVersion hie <- readHieFileCompat ref path
+  when (GHC.hieVersion /= fileHieVersion) $ do
+    putStrLn $ "WARNING: version mismatch in " <> path
+    putStrLn $ "    The hie files in this project were generated with GHC version: " <> show fileGHCVersion
+    putStrLn $ "    This version of calligraphy was compiled with GHC version: " <> show GHC.hieVersion
+    putStrLn "    Optimistically continuing anyway..."
+  pure hie
+
+
 {-# INLINE sourceInfo #-}
 sourceInfo :: Traversal' (HieAST a) (NodeInfo a)
 showContextInfo :: ContextInfo -> String
 readHieFileCompat :: IORef NameCache -> FilePath -> IO HieFileResult
-#if MIN_VERSION_ghc(9,0,0)
+
+#if MIN_VERSION_ghc(9,4,0)
+
+sourceInfo f (Node (SourcedNodeInfo inf) sp children) = (\inf' -> Node (SourcedNodeInfo inf') sp children) <$> Map.alterF (maybe (pure Nothing) (fmap Just . f)) SourceInfo inf
+
+showContextInfo = showSDocUnsafe . ppr
+
+readHieFileCompat ref path = do
+  nameCache <- readIORef ref
+  readHieFile nameCache path
+
+#elif MIN_VERSION_ghc(9,0,0)
 
 sourceInfo f (Node (SourcedNodeInfo inf) sp children) = (\inf' -> Node (SourcedNodeInfo inf') sp children) <$> Map.alterF (maybe (pure Nothing) (fmap Just . f)) SourceInfo inf
